+++ Makefile
@@ -8,7 +8,7 @@ all: $(CLIENT) $(SERVER)
 CFLAGS := -Wall -g -O2
 LDFLAGS := -ludev -lyaml
 
-CLIENT_SRCS := cdba.c circ_buf.c
+CLIENT_SRCS := cdba.c cdba-client.c circ_buf.c
 CLIENT_OBJS := $(CLIENT_SRCS:.c=.o)
 
 SERVER_SRCS := cdba-server.c cdb_assist.c circ_buf.c conmux.c device.c device_parser.c fastboot.c alpaca.c console.c qcomlt_dbg.c
